+++ pylate/evaluation/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from .beir import evaluate, get_beir_triples, load_beir
-from .code_stack_network_evaluator import CodeSearchNetworkEvaluator
+from .code_search_network_evaluator import CodeSearchNetworkEvaluator
 from .colbert_distillation import ColBERTDistillationEvaluator
 from .colbert_triplet import ColBERTTripletEvaluator
 from .custom_dataset import load_custom_dataset
